+++ lib/components/touch_scale_context.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/animation.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter_touch_scale/components/touch_scale_behavior.dart';
 import 'package:flutter_touch_scale/flutter_touch_scale.dart';
 
 @protected
@@ -11,4 +12,5 @@ mixin TouchScaleContext {
   Curve get reverseCurve;
   Duration? get previewDuration;
   TouchScaleCallPhase get callPhase;
+  TouchScaleBehavior get behavior;
 }