+++ lib/flutter_touch_scale.dart
@@ -7,3 +7,4 @@ export 'widgets/touch_scale_gesture_detector.dart';
 export 'components/touch_scale_context.dart';
 export 'components/touch_scale_controller.dart';
 export 'components/touch_scale_gesture_recognizer.dart';
+export 'components/touch_scale_behavior.dart';